Moved PropSpec related tests to propspec-test, please note that currently AsyncPropSpecXXXSpec.scala (from org.scalatest.fixture) are currently all commentted out.
