+++ build.sbt
@@ -21,6 +21,7 @@ lazy val scalatestFlatSpecTest = ScalatestBuild.scalatestFlatSpecTest
 lazy val scalatestFreeSpecTest = ScalatestBuild.scalatestFreeSpecTest
 lazy val scalatestFunSpecTest = ScalatestBuild.scalatestFunSpecTest
 lazy val scalatestFunSuiteTest = ScalatestBuild.scalatestFunSuiteTest
+lazy val scalatestPropSpecTest = ScalatestBuild.scalatestPropSpecTest
 lazy val scalatestJS = ScalatestBuild.scalatestJS
 lazy val scalatestNative = ScalatestBuild.scalatestNative
 lazy val scalatestDotty = ScalatestBuild.scalatestDotty
