+++ jvm/propspec-test/src/test/scala/org/scalatest/propspec/AsyncPropSpecSpec2.scala
